Tf2 core code is not dependent on ROS. This is compiled as a static libray, this code is taken from https://github.com/ros2/geometry2 tf2 module. Small changes such as the ROS messages have been written in tf2/include/msg and instead of ros::time used the local time.hpp is used.
Many changes:
* Tf2 core code is not dependent on ROS. This is compiled as a static libray, this code is taken from
https://github.com/ros2/geometry2 tf2 module. Small changes such as the ROS messages have been written in
tf2/include/msg and instead of ros::time used the local time.hpp is used.

* Most of the mapper module is done, mapper.cpp does not have the `main loop` done yet. To work propely it needs the
needs information from the tracker and therefore the messaging system between them is needed. But when that's fixed
most the rest should be easy. Nothing very noteworthy here, mosly copied code from ES-PTAM with some small modifications.

* Started implementing the tracker. certain files could just be copied over with small changes, other need more work. The tracker.cpp will require some work,
but when the messaging between the mapper and tracker works it should not be too difficult to finish it.

* Some includes are from another direcotry evo_utils, have not fixed that yet.

* Changed the structre, so there is app, with a main.cpp, CMakeLists.txt and a include dir. Then there is a mapper and tracker dir.