Mapper can now recive and set transform frame, the code from ES-PTAM has just been copied and modfied to work here.
Fixed some things in generall, should do a commit for 1 thing because now I dont remember everything that is changed.

It's not done yet, beacuse the tracker still needs a camera but should be easy to fix.

Also, it is a good idea to fix some cv images for debbuging as well. And how to strem the global PCL!